test(parser): add parser and divide parser logic into multiple directory
