+++ src/Error/handleError.js
@@ -1,4 +1,4 @@
 export default function handleError(error) {
   // [ERROR]로 시작하는 에러 메시지를 출력
-  console.error(`[ERROR] ${error.message}`);
+  console.error(`[ERROR]`);
 }
